fix sickies on overscroll effect

+++ app/src/main/kotlin/app/atomofiron/searchboxapp/custom/view/ExplorerStickyTopView.kt
@@ -58,8 +58,8 @@ class ExplorerStickyTopView(
         composition?.let { binder.bindComposition(it) }
     }
 
-    fun drawTop(size: Int) {
-        drawTop = size.toFloat()
+    fun drawTop(size: Float) {
+        drawTop = size
         invalidate()
     }
 }
